Code Smell revision (StaffRating) [UC-31] [UC-32] [Implementation] [refs issue #63] [refs issue #66]

+++ src/main/java/lapr/project/utils/StaffRating.java
@@ -131,7 +131,7 @@ public class StaffRating {
 
         for (Application app : event.getApplicationRegister().getApplicationList()) {
             for (Review rev : app.getListReview()) {
-                double rmean = rev.getStaffTopicKnowledge() + rev.getEventAdequacy() + rev.getInviteAdequacy() + rev.getAreaAdequacy() + rev.getRecommendation();
+                double rmean = (double) rev.getStaffTopicKnowledge() + rev.getEventAdequacy() + rev.getInviteAdequacy() + rev.getAreaAdequacy() + rev.getRecommendation();
                 rmean = rmean / 5;
                 rsum += rmean;
                 rcount++;
